requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.1
dataclasses-json==0.5.14
distro==1.9.0
etelemetry==0.3.1
faiss-cpu==1.7.4
filelock==3.13.1
fitz==0.0.1.dev2
frontend==0.0.3
frozenlist==1.4.1
fsspec==2023.12.2
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.20.3
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.295
langchain-community==0.0.24
langchain-core==0.1.26
langsmith==0.0.92
llvmlite==0.41.1
looseversion==1.3.0
lxml==5.1.0
MarkupSafe==2.1.3
marshmallow==3.20.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
numba==0.58.1
numexpr==2.9.0
numpy==1.26.2
openai==0.27.6
openai-whisper==20231117
orjson==3.9.15
packaging==23.2
pandas==2.2.1
pathlib==1.0.1
pillow==10.2.0
pip==23.3.2
prov==2.0.0
pydantic==2.6.2
pydantic_core==2.16.3
pydot==2.0.0
PyMuPDF==1.23.25
PyMuPDFb==1.23.22
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==7.0.0
regex==2023.10.3
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.4.0
setuptools==65.5.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.37.1
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tiktoken==0.5.2
tokenizers==0.15.2
torch==2.1.2
tqdm==4.66.1
traits==6.3.2
transformers==4.38.1
typing_extensions==4.9.0
typing-inspect==0.9.0
tzdata==2024.1
urllib3==2.1.0
uvicorn==0.27.1
yarl==1.9.4